requirements.txt
tensorflow
opencv-python
mediapipe
numpy
streamlit
playsound==1.2.2
pyyaml
scikit-learn
matplotlib
